server/requirements.txt
fastmcp
msal
PyJWT
pydantic-settings
python-dotenv
aiohttp
microsoft-agents-activity
microsoft-agents-copilotstudio-client